requirements.txt
pymssql
python-dotenv